// apps/exam-vault/src/generated/models/question-model.ts
import type { Section } from './section-model';

export const QuestionQuestionTypeKeyToLabel = {
  'QuestionTypeKey0': 'MCQ',
  'QuestionTypeKey1': 'ShortText',
  'QuestionTypeKey2': 'Code',
  'QuestionTypeKey3': 'Listening',
  'QuestionTypeKey4': 'Reading',
  'QuestionTypeKey5': 'Writing',
  'QuestionTypeKey6': 'Speaking',
  'QuestionTypeKey7': 'Interview'
} as const;
export type QuestionQuestionTypeKey = keyof typeof QuestionQuestionTypeKeyToLabel;

export interface Question {
  /**
   * @displayName Question ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Question Title
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  questionTitle: string;
  /**
   * @displayName Code Language
   */
  codeLanguage?: string;
  /**
   * @displayName Correct Answer
   */
  correctAnswer?: string;
  /**
   * @displayName Keywords
   */
  keywords?: string;
  /**
   * @displayName Options (JSON)
   */
  optionsJSON?: string;
  /**
   * @displayName Order Index
   * @validationRule Required for create/update operations
   */
  orderIndex: number;
  /**
   * @displayName Points
   * @validationRule Required for create/update operations
   */
  points: number;
  /**
   * @displayName Question Text
   * @validationRule Required for create/update operations
   */
  questionText: string;
  /**
   * @displayName Question Type
   * @validationRule Required for create/update operations
   */
  questionTypeKey: QuestionQuestionTypeKey;
  /**
   * @displayName Section
   * @validationRule Required for create/update operations
   */
  section: Pick<Section, 'id' | 'name1'>;
}

export const _Question = 'Question' as const;